Added Reader Writer to Copy tar
A program to implement Reader-Writer problem of OS. Done to implement copying of a TAR file , then validation testing is done using the md5sum check. md5sum is sent by reader to writer which then calculated md5sum of file and sent it to reader via buffer. Reader then calculates md5sum of received file and compares it to the received sum to confirm and validate the file received.